pyproject.toml
[project]
name = "agents-with-strands"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.68.1",
    "fastapi>=0.117.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]
